china_border/requirements.txt
APScheduler==3.0.5
Babel==2.3.4
backports-abc==0.4
bcrypt==3.1.1
blinker==1.4
cffi==1.8.3
colorama==0.3.7
cssmin==0.2.0
defusedxml==0.4.1
feedparser==5.2.1
Flask==0.10.1
#Flask-Admin==1.4.2
Flask-Assets==0.12
Flask-Babel==0.11.1
#Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Login==0.3.2
flask-login-openerp==0.2.0
Flask-Mail==0.9.1
flask-mongoengine==0.7.0
Flask-OpenERP==0.3.1
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-QRcode==0.6.0
Flask-Restless==0.17.0
Flask-Script==2.0.5
Flask-SocketIO==2.7.1
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
#flask-themes==0.1.4
#Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
jsmin==2.2.1
Markdown==2.6.4
MarkupSafe==0.23
mimerender==0.6.0
mongoengine==0.10.6
nose==1.3.7
numpy==1.11.1
Pillow==3.1.1
psycopg2==2.6.1
pycparser==2.16
pycrypto==2.6.1
#pycurl==7.43.0
PyExecJS==1.4.0
Pygments==2.0.2
pymongo==3.3.0
PyMySQL==0.6.7
pypinyin==0.9.1
PyReact==0.6.0
python-dateutil==2.5.3
python-engineio==1.0.3
python-mimeparse==1.6.0
python-socketio==1.5.1
python3-openid==3.0.10
pytz==2015.7
qrcode==5.2.2
requests==2.9.1
six==1.10.0
SQLAlchemy==1.0.11
SQLAlchemy-Utils==0.32.9
suds-jurko==0.6
tornado==4.3
tzlocal==1.2
Unidecode==0.4.19
webassets==0.12.0
wechat==0.4.16
Werkzeug==0.11.3
#WTForms==2.1
